[WIP] more changes to fix some previous issues merged in PRs and naming updates
